Refactor NewEventPage WIP
backend functionalities not done
