+++ src/assets/styles.css
@@ -1,0 +1,8 @@
+.form-control:valid {
+  background-image: none !important;
+  border-color: #ced4da !important;
+}
+
+.form-check-input:valid ~ .form-check-label {
+  color: #212529 !important;
+}
