+++ src/index.tsx
@@ -11,7 +11,7 @@ import EventSignup from './components/EventSignup'
 import Login from './components/Login'
 import EventManagement from './components/EventManagement'
 import AdminPage from './components/AdminPage'
-import NewEvent from './components/NewEvent'
+import { NewEvent } from './components/NewEvent'
 import Logout from './components/Logout'
 import IndexRoute from './components/IndexRoute'
 import { KeycloakProvider } from './auth/keycloak'
